# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Management System API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/branch": {
            "get": {
                "tags": [
                    "branch"
                ],
                "summary": "Get all branches",
                "operationId": "7a60d8d8fd907b42bfb98cf328e9b958",
                "responses": {
                    "200": {
                        "description": "List of all branches",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Main Branch"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "Banasree"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-05-05 12:00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2021-05-05 12:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "branch"
                ],
                "summary": "Create new branch",
                "description": "Create new branch and return created data",
                "operationId": "createBranch",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "main-branch"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Dhaka"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "Branch ID",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "description": "Branch name",
                                                    "type": "string",
                                                    "example": "main-branch"
                                                },
                                                "location": {
                                                    "description": "Branch location",
                                                    "type": "string",
                                                    "example": "Dhaka"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2021-05-05 12:00:00"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2021-05-05 12:00:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}